fix: qr 링크 변경 및 isApple util 함수 변경
