+++ src/pages/index.tsx
@@ -40,6 +40,7 @@ import {
   INSTALL_APP_EXP,
 } from "@/constants/Application";
 import { queryClient } from "@/service/react-query/queryClient";
+import isApple from "@/util/isApple"
 
 const Home = () => {
   const theme = useTheme();
@@ -89,7 +90,7 @@ const Home = () => {
         qr: {
           text: "QR 코드 스캔하고 설치",
           qrValues: {
-            value: "https://pink1016.tistory.com/",
+            value: isApple() ? APP_STORE : PLAY_STORE,
             size: 60,
 
           }
@@ -123,32 +124,6 @@ const Home = () => {
     const expires = getPopUpExpireDate();
     setCookie("BANNER_EXPIRES", true, { path: "/", expires });
     setHasCookie(true);
-  };
-
-  const isApple = () => {
-    const user = navigator.userAgent.toLowerCase();
-    if (!isMobile) {
-      if (
-        user.indexOf("apple") > -1 ||
-        user.indexOf("safari") > -1 ||
-        user.indexOf("mac") > -1
-      ) {
-        return true;
-      } else {
-        return false;
-      }
-    } else {
-      if (
-        user.indexOf("iphone") > -1 ||
-        user.indexOf("ipad") > -1 ||
-        user.indexOf("ipod") > -1 ||
-        user.indexOf("safari") > -1
-      ) {
-        return true;
-      } else {
-        return false;
-      }
-    }
   };
 
   const onInstallApp = () => {
